Reorganised the file structure and working on adding  bottom navigation bar on the pages
